Fallout of a hacking Sunday. Tackles  #7567 and #8417.

Tested against my own project, only, as I'm not aware how the testsuite works.

Not so sure about commit 0d91578, _"don't shortcut on previously seen elements"_, as I'm not sure how much performance impact it has. It solves handling ignored tags, but probably needs either removal of then dead code or a different solution for dealing with ignored tags appearing multiple times.