@Traumflug 2021-04-19
In case somebody wants to enhance the testsuite, here are additional cases:
```
<script type="text/javascript">
    for (var i = 0; i < imgDefer.length; i++) {
        doNothing();
    }
</script>
```
```
<script
  type="text/javascript"
>
  if (a<b && a>c) {
    haha();
  }
</script>
```
```
<form
	action="[URL]"
	method="post"
>
```
```
<hr	id=a>
```
Yes, newlines and tab vs. space matters :-)

Failure of each is recognized by non-alphabetical characters appearing in the `hugo_stats.json`.