@bep 2021-04-20
>Tested against my own project, only, as I'm not aware how the testsuite works.

```
go test ./publisher
```