@bep 2021-04-20
Thanks for this.  

There are failing tests. It looks mostly to be elements inside pre/textarea, which I don't see any reason to collect (but please prove me wrong).

Also, it would be great if you could 

1. Add some test cases for the cases this PR is supposed to fix.
2. Run the benchmark inside the ./publisher package and compare it to master (see below) 


```bash
name                     old time/op    new time/op    delta
ClassCollectorWriter-16    21.2µs ± 2%    47.3µs ± 1%  +123.21%  (p=0.029 n=4+4)

name                     old alloc/op   new alloc/op   delta
ClassCollectorWriter-16    35.3kB ± 0%    86.4kB ± 0%  +144.95%  (p=0.029 n=4+4)

name                     old allocs/op  new allocs/op  delta
ClassCollectorWriter-16       155 ± 0%       329 ± 0%  +112.26%  (p=0.029 n=4+4)
```

Note that I don't mind it getting slower if really needed, but I fail to understand the motivation behind the removal of `elementSet`, which I suspect is the reason why it is now so much slower and memory hungry.