@Traumflug 2021-04-22
Thanks for the hint on how to run these testcases easily. I just pushed a reviewed set of commits.

> There are failing tests. It looks mostly to be elements inside pre/textarea, which I don't see any reason to collect

Well, the simple reason for these failures was me removing special handling of pre/textarea, but not knowing how to remove these tests as well.

That said, removing these was a mistake, I misunderstood the intention of this code earlier. It's not about not collecting non-styleable tags, but about ignoring stuff inside them. Accordingly, I melted down the first commit to just adding another two such tags with preformatted code. Where _<title>_ isn't exactly preformatted, but of no use for styling anyways.

> Can you check whether the latest commits re this in the master branch covers your needs?

Well, some, but not all. This time, knowing how to deal with these tests, I managed to find a testcase for each. Committed testcases separately to allow seeing them failing. Luckily, I also found a fix for each.

Also one or two testcases which fell into the keyboard somehow, without a matching reported issue I'd be aware of :-)

> Note that I don't mind it getting slower if really needed, but I fail to understand the motivation behind the removal of elementSet

This time I moved the commit with this removal, along with a commit with a demonstrating (previously failing, now working) testcase just before it, all to the end to allow (cherry-)picking earlier commits.

The problem to solve appears when such a preformatted tag appears twice and contains a `<` (happens rarely in testcases, almost always in real world usage). First round everything is fine and dandy. On the second round, the processing time saver misses the part where `w.inPreTag` gets set. Accordingly, the following parsing doesn't recognize the preformatted tag as such and treats it like a normal tag.

Perhaps you have a better idea on how to get the best of both, quick processing as well as recognizing repeated preformatted tags.